Do not include `vendor/bundle` in GHA built PRs (#75)
* Only run CI when `docs/src` changed

https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#onpushpull_requestpull_request_targetpathspaths-ignore

* Include language in PR title

* Ignore GHA generated Ruby files

The PRs generated by the GHA have some 400,000+ lines changed in them https://github.com/heroku/buildpacks/pull/71. That is because it 2,000+ files installed including all the gems installed and Ruby. This is definitively not what we want. This behavior is documented in https://github.com/ruby/setup-ruby/tree/v1/?tab=readme-ov-file#bundle-config that it will place the files in `vendor/bundle`. To exclude the directory, we can add it to our `.gitignore`.