Update PR tasks
- Add only the docs in docs/fail to one PR and send it with an updated message and branch
- Add only the docs not in docs/fail to another pr and send it as before

## References

This task uses a github action:

https://github.com/peter-evans/create-pull-request/tree/v7/?tab=readme-ov-file#ignore-files

Here's a local demo of this git add syntax

```
$ echo "good change" >> docs/ruby/README.md
$ mkdir -p docs/fail/ruby && echo "failed" >> docs/fail/ruby/README.md
$ git status
On branch main
Your branch is up to date with 'origin/main'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   docs/ruby/README.md

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	docs/fail/

no changes added to commit (use "git add" and/or "git commit -a")
$ git add -- docs/* ':!docs/fail/*'
$ git status
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	modified:   docs/ruby/README.md

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	docs/fail/

$ git add docs/fail/*
$ git status
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
	new file:   docs/fail/ruby/README.md
	modified:   docs/ruby/README.md
```

I'm unsure if this is the right syntax for the action for `add-paths`.